Moved the inspection logic to its own file
